tests: move ip bit operations tests to seperate file
The getbits tests don't belong to lpfst tree implementation. If the
lpfst tree is removed or changed the getbit tests can stay as they are.
Outsource them to a separate file to make them independent from the
lpfst tests.
