+++ CMakeLists.txt
@@ -85,6 +85,8 @@ ADD_TEST(test_live_validation tests/test_live_validation)
 
 ADD_TEST(test_ipaddr tests/test_ipaddr)
 
+ADD_TEST(test_getbits tests/test_getbits)
+
 #install lib
 set (RTRLIB_VERSION_MAJOR 0)
 set (RTRLIB_VERSION_MINOR 3)
